index dane

+++ inlämingbank/BankAppData/Account.cs
@@ -16,7 +16,7 @@ namespace inlämingbank.BankAppData
         public int AccountId { get; set; }
         public string Frequency { get; set; } = null!;
         public DateTime Created { get; set; }
-        public decimal Balance { get; set; }
+        public int Balance { get; set; }
 
         public virtual ICollection<Disposition> Dispositions { get; set; }
         public virtual ICollection<Loan> Loans { get; set; }
